Massive Template Update
Unfortunately, it was difficult to separate this commit into smaller commits do the linked nature of all the files. Therefore, it is one large commit.

Currently Normalize.cpp and Grid2D.h both utilize functions that are not template-able. Sparse Matrices do not have functionality to do “.each_row()” or to do .rows(vector_of_indices). Therefore, we make two overloaded functions to redo this. However, it would be better to replace the functionality with an inline function in utils.h that is template-able.
