+++ src/MakeCD.cpp
@@ -1,31 +1,31 @@
 #include "MakeCD.h"
 
-
-CD * make_CD(const arma::mat& Xi, const arma::vec& yi, const Params& P)
+template <typename T>
+CD<T> * make_CD(const T& Xi, const arma::vec& yi, const Params<T>& P)
 {
 
     if (P.Specs.SquaredError)
     {
         if (P.Specs.CD)
         {
-            if (P.Specs.L0) {return new CDL0(Xi, yi, P);}
-            else {return new CDL012(Xi, yi, P);}
+            if (P.Specs.L0) {return new CDL0<T>(Xi, yi, P);}
+            else {return new CDL012<T>(Xi, yi, P);}
         }
-        else if (P.Specs.PSI) {return new CDL012Swaps(Xi, yi, P);}
+        else if (P.Specs.PSI) {return new CDL012Swaps<T>(Xi, yi, P);}
     }
 
     else if (P.Specs.Logistic)
     {
-        if (P.Specs.CD) {return new CDL012Logistic(Xi, yi, P);}
-        else if (P.Specs.PSI) {return new CDL012LogisticSwaps(Xi, yi, P);}
+        if (P.Specs.CD) {return new CDL012Logistic<T>(Xi, yi, P);}
+        else if (P.Specs.PSI) {return new CDL012LogisticSwaps<T>(Xi, yi, P);}
     }
 
     else if (P.Specs.SquaredHinge)
     {
-        if (P.Specs.CD) {return new CDL012SquaredHinge(Xi, yi, P);}
-        else if (P.Specs.PSI) {return new CDL012SquaredHingeSwaps(Xi, yi, P);}
+        if (P.Specs.CD) {return new CDL012SquaredHinge<T>(Xi, yi, P);}
+        else if (P.Specs.PSI) {return new CDL012SquaredHingeSwaps<T>(Xi, yi, P);}
     }
 
-    return new CDL0(Xi, yi, P); // handle later
+    return new CDL0<T>(Xi, yi, P); // handle later
 
 }
