+++ src/RInterface.cpp
@@ -1,9 +1,10 @@
 #include "RcppArmadillo.h"
 #include "Grid.h"
 // [[Rcpp::depends(RcppArmadillo)]]
+// [[Rcpp::plugins(cpp11)]]
 
-// [[Rcpp::export]]
-Rcpp::List L0LearnFit(const arma::mat& X, const arma::vec& y, const std::string Loss, const std::string Penalty,
+template <typename T>
+Rcpp::List _L0LearnFit(const T& X, const arma::vec& y, const std::string Loss, const std::string Penalty,
                       const std::string Algorithm, const unsigned int NnzStopNum, const unsigned int G_ncols, const unsigned int G_nrows,
                       const double Lambda2Max, const double Lambda2Min, const bool PartialSort,
                       const unsigned int MaxIters, const double Tol, const bool ActiveSet,
@@ -11,7 +12,7 @@ Rcpp::List L0LearnFit(const arma::mat& X, const arma::vec& y, const std::string 
                       const double ScaleDownFactor, unsigned int ScreenSize, const bool LambdaU, const std::vector< std::vector<double> > Lambdas, const unsigned int ExcludeFirstK, const bool Intercept)
 {
     auto p = X.n_cols;
-    GridParams PG;
+    GridParams<T> PG;
     PG.NnzStopNum = NnzStopNum;
     PG.G_ncols = G_ncols;
     PG.G_nrows = G_nrows;
@@ -24,7 +25,7 @@ Rcpp::List L0LearnFit(const arma::mat& X, const arma::vec& y, const std::string 
     PG.LambdasGrid = Lambdas;
     PG.Lambdas = Lambdas[0]; // to handle the case of L0 (i.e., Grid1D)
     PG.intercept = Intercept;
-    Params P;
+    Params<T> P;
     P.MaxIters = MaxIters;
     P.Tol = Tol;
     P.ActiveSet = ActiveSet;
@@ -48,7 +49,7 @@ Rcpp::List L0LearnFit(const arma::mat& X, const arma::vec& y, const std::string 
     //case "L1": PG.P.Specs.L1 = true;
     //case "L1Relaxed": PG.P.Specs.L1Relaxed = true;
 
-    Grid G(X, y, PG);
+    Grid<T> G(X, y, PG);
     G.Fit();
 
     std::string FirstParameter = "lambda";
@@ -115,4 +116,38 @@ Rcpp::List L0LearnFit(const arma::mat& X, const arma::vec& y, const std::string 
 
     return l;
 
+}
+
+// [[Rcpp::export]]
+Rcpp::List L0LearnFit(const SEXP& X, const arma::vec& y, const std::string Loss, const std::string Penalty,
+                      const std::string Algorithm, const unsigned int NnzStopNum, const unsigned int G_ncols, const unsigned int G_nrows,
+                      const double Lambda2Max, const double Lambda2Min, const bool PartialSort,
+                      const unsigned int MaxIters, const double Tol, const bool ActiveSet,
+                      const unsigned int ActiveSetNum, const unsigned int MaxNumSwaps,
+                      const double ScaleDownFactor, unsigned int ScreenSize, const bool LambdaU, const std::vector< std::vector<double> > Lambdas, const unsigned int ExcludeFirstK, const bool Intercept){
+if (Rf_isS4(X) && Rf_inherits(X, "dgCMatrix")){
+    arma::sp_mat m = Rcpp::as<arma::sp_mat>(X);
+    return _L0LearnFit(m, y, Loss, Penalty,
+                       Algorithm, NnzStopNum, G_ncols, G_nrows,
+                       Lambda2Max, Lambda2Min, PartialSort,
+                       MaxIters, Tol, ActiveSet,
+                       ActiveSetNum, MaxNumSwaps,
+                       ScaleDownFactor, ScreenSize, LambdaU, 
+                       Lambdas, ExcludeFirstK, Intercept);
+  }
+  else if (Rf_isS4(X) && Rf_inherits(X, "dgeMatrix"))
+  {
+      Rcpp::stop("Only supports Dense and dgCMatrix Matricies");
+  }
+  else
+  {
+    arma::mat m = Rcpp::as<arma::mat>(X);
+    return _L0LearnFit(m, y, Loss, Penalty,
+                       Algorithm, NnzStopNum, G_ncols, G_nrows,
+                       Lambda2Max, Lambda2Min, PartialSort,
+                       MaxIters, Tol, ActiveSet,
+                       ActiveSetNum, MaxNumSwaps,
+                       ScaleDownFactor, ScreenSize, LambdaU, 
+                       Lambdas, ExcludeFirstK, Intercept);
+  }
 }
