+++ src/include/Params.h
@@ -5,6 +5,7 @@
 #include "RcppArmadillo.h"
 #include "Model.h"
 
+template <typename T>
 struct Params
 {
 
@@ -28,9 +29,12 @@ struct Params
     unsigned int Iter = 0; // Current iteration number in the grid
     unsigned int ScreenSize = 1000;
     arma::vec * r;
-    arma::mat * Xy; // used for classification.
+    T * Xy; // used for classification.
     unsigned int NoSelectK = 0;
     bool intercept;
 };
+
+template struct Params<arma::mat>;
+template struct Params<arma::sp_mat>;
 
 #endif
